refactor: APIのエラーレスポンスとクライアントのエラーオブジェクトをマッピングする処理を切り出した
