+++ src/resources/ts/http/client.ts
@@ -1,21 +1,16 @@
-import { ValidationError } from "@/ts/error/ValidationError";
-import { validationErrorType } from "@/ts/http/errorResponseType";
+import {createError} from "@/ts/http/error/errorFactory";
 
 export const client = (input: RequestInfo | URL, init?: RequestInit): Promise<Response> => (
     fetch(input, init)
         .then(async (data) => {
             if (data.ok) return data.json()
 
-            const json = await data.json()
-            if (!json.type) throw new Error(json.message)
-
-            switch (json.type) {
-                case validationErrorType:
-                    const userMessages = pickUserMessages(json.user_messages);
-                    throw new ValidationError(json.message, userMessages);
-                default:
-                    throw new Error('Unexpected error type.');
+            const error = await data.json()
+            if (error.type) {
+                throw createError(error);
             }
+
+            throw new Error(error.message)
         })
 );
 
@@ -24,18 +19,3 @@ export const getCsrfToken = (): string | null | undefined => {
         .querySelector('meta[name="csrf-token"]')
         ?.getAttribute('content');
 };
-
-/**
- * フォームごとに複数あるエラーメッセージを1つに変換する
- *
- * Laravelはフォームごとに複数のエラーメッセージを返すが、Material-UIのhelperTextは1つのエラーメッセージしか表示できない。
- * 複数のエラーメッセージを同時に表示したいケースも多くないので、エラーメッセージは1つずつ表示する。
- */
-const pickUserMessages = (userMessages: Record<string, any>): Record<string, string> => {
-    const pickedMessages: Record<string, string> = {};
-    for (const key in userMessages) {
-        pickedMessages[key] = userMessages[key][0];
-    }
-
-    return pickedMessages;
-}
