+++ src/resources/ts/http/error/errorResponseType.ts
